requirements.txt
faiss-cpu
numpy
cachetools
asyncio
llama-index
llama-parse
streamlit
loguru
pyyaml
google-auth
python-dotenv
PyMuPDF
openai
tf-keras
tensorflow-model-optimization
pinecone-client
llama-index-llms-azure-openai
llama_index.embeddings.azure_openai
llama_index.embeddings.huggingface
llama_index.llms.huggingface
llama_index.llms.vertex
llama_index.embeddings.vertex
llama_index.llms.gemini
llama_index.vector_stores.pinecone
llama-index-llms-groq
"anthropic[vertex]"